test: delete ten tests that asserted against their own mock (#378)
tests/fuse_test.go built a fuse.FileSystem in each of its three top-level
functions and then wrote `_ = filesystem // Use filesystem to avoid unused
variable`. Of its ten subtests exactly one touched it, reading GetStats() on
either side of a loop that called backend.GetObject directly — so
assert.GreaterOrEqual(statsAfter.Reads, statsBefore.Reads) compared 0 against 0,
which holds for every possible implementation of internal/fuse including one
whose Read panics. The rest asserted that an in-memory map returns what was put
into it, and BenchmarkFUSEOperations measured that map's throughput and reported
it as FUSE performance.

The file survives as tests/mockbackend_test.go carrying only the MockBackend
fixture, which integration_test.go, unit_test.go, and predictive_cache_test.go
need at six call sites. Everything below the fixture is deleted, which also
removes internal/fuse from the tests package's import graph — it was the only
importer there.

Checked per subtest rather than assumed: write coalescing is asserted harder by
internal/vfs/writer_test.go's forty tests, the cache subtest duplicates part of
internal/cache/multilevel_test.go's twenty-three, metrics has four test files of
its own, and the read-ahead claim is what internal/fuse/read_path_test.go
actually makes across fourteen tests driving the real FileHandle.Read.
internal/fuse coverage is unchanged at 68.6% against a floor of 67.

Verified by mutation, since the point is that these tests could not fail:
zeroing Reads in GetStats fails TestGetStatsReportsEveryCounter by name, and
forcing FileHandle.Read's offset to 0 fails seven read-path tests by name. Under
both mutations `go test -race ./tests/` passed exactly as before the deletion.

Two stale references fixed alongside: the four PutObject test-double line numbers
in docs/design/conditional-writes-vs-raft.md had all drifted, not only the
renamed one, and the v0.13.0 entry describing the defer-to-t.Cleanup move now
notes that its example file is gone while the finding stands.
