--- tests/fuse_test.go
@@ -1,554 +1,0 @@
-package tests
-
-import (
-	"context"
-	"maps"
-	"os"
-	"strings"
-	"sync"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/scttfrdmn/objectfs/internal/cache"
-	"github.com/scttfrdmn/objectfs/internal/fuse"
-	"github.com/scttfrdmn/objectfs/internal/metrics"
-	"github.com/scttfrdmn/objectfs/internal/vfs"
-	"github.com/scttfrdmn/objectfs/pkg/types"
-)
-
-// MockBackend implements a simple in-memory backend for testing
-type MockBackend struct {
-	mu      sync.RWMutex
-	objects map[string][]byte
-
-	// meta holds each object's user metadata. Stored rather than discarded because that is where POSIX
-	// mode and ownership live: a backend that accepted attributes and dropped them would let an
-	// attribute test pass while nothing was persisted.
-	meta map[string]map[string]string
-}
-
-func NewMockBackend() *MockBackend {
-	return &MockBackend{
-		objects: make(map[string][]byte),
-		meta:    make(map[string]map[string]string),
-	}
-}
-
-func (b *MockBackend) GetObject(ctx context.Context, key string, offset, size int64) ([]byte, error) {
-	b.mu.RLock()
-	defer b.mu.RUnlock()
-
-	data, exists := b.objects[key]
-	if !exists {
-		return nil, os.ErrNotExist
-	}
-
-	if offset >= int64(len(data)) {
-		return []byte{}, nil
-	}
-
-	end := offset + size
-	if size == 0 || end > int64(len(data)) {
-		end = int64(len(data))
-	}
-
-	return data[offset:end], nil
-}
-
-func (b *MockBackend) PutObject(ctx context.Context, key string, data []byte, meta map[string]string) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-
-	b.objects[key] = make([]byte, len(data))
-	copy(b.objects[key], data)
-	// Replaced wholesale, as S3's PutObject does.
-	b.meta[key] = copyMeta(meta)
-	return nil
-}
-
-// PutObjectIf refuses rather than pretending. This mock could evaluate a precondition against its own
-// map, but an in-process map is not what a precondition is a claim about — the point of one is that a
-// *remote* store arbitrates between writers this process cannot see. Implementing it here would produce
-// a CAS that passes its tests and cannot exclude anything, which is the seam-agreement failure the
-// project uses internal/testaws to avoid. Conditional writes are tested against a real S3 endpoint.
-func (b *MockBackend) PutObjectIf(ctx context.Context, key string, data []byte, meta map[string]string,
-	cond types.Precondition,
-) (string, error) {
-	return "", types.ErrNotSupported
-}
-
-func (b *MockBackend) SetObjectMetadata(ctx context.Context, key string, meta map[string]string) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-
-	if _, ok := b.objects[key]; !ok {
-		return os.ErrNotExist
-	}
-	// The bytes are untouched. An implementation that rewrote them would be the defect
-	// vfs.Flusher asserts against by rechecking the size after an attribute write.
-	b.meta[key] = copyMeta(meta)
-	return nil
-}
-
-// copyMeta returns a copy of m, so a caller reusing its map cannot mutate what the backend stored.
-func copyMeta(m map[string]string) map[string]string {
-	out := make(map[string]string, len(m))
-	maps.Copy(out, m)
-	return out
-}
-
-func (b *MockBackend) HeadObject(ctx context.Context, key string) (*types.ObjectInfo, error) {
-	b.mu.RLock()
-	defer b.mu.RUnlock()
-
-	data, exists := b.objects[key]
-	if !exists {
-		return nil, os.ErrNotExist
-	}
-
-	return &types.ObjectInfo{
-		Key:          key,
-		Size:         int64(len(data)),
-		LastModified: time.Now(),
-		Metadata:     copyMeta(b.meta[key]),
-	}, nil
-}
-
-func (b *MockBackend) ListObjects(ctx context.Context, prefix string, maxKeys int) ([]types.ObjectInfo, error) {
-	b.mu.RLock()
-	defer b.mu.RUnlock()
-
-	var objects []types.ObjectInfo
-	count := 0
-
-	for key, data := range b.objects {
-		if (prefix == "" || strings.HasPrefix(key, prefix)) && count < maxKeys {
-			objects = append(objects, types.ObjectInfo{
-				Key:          key,
-				Size:         int64(len(data)),
-				LastModified: time.Now(),
-			})
-			count++
-		}
-	}
-
-	return objects, nil
-}
-
-// CopyObject copies the bytes and the metadata together, which is the property rename depends on: the
-// stored mode, ownership, and content encoding have to arrive at the destination or the renamed file
-// comes back with a different mode, or — if it was compressed — unreadable.
-func (b *MockBackend) CopyObject(ctx context.Context, src, dst string) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-
-	data, exists := b.objects[src]
-	if !exists {
-		return os.ErrNotExist
-	}
-
-	b.objects[dst] = append([]byte(nil), data...)
-	b.meta[dst] = copyMeta(b.meta[src])
-	return nil
-}
-
-func (b *MockBackend) DeleteObject(ctx context.Context, key string) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-
-	delete(b.objects, key)
-	return nil
-}
-
-func (b *MockBackend) GetObjects(ctx context.Context, keys []string) (map[string][]byte, error) {
-	b.mu.RLock()
-	defer b.mu.RUnlock()
-
-	result := make(map[string][]byte)
-	for _, key := range keys {
-		if data, exists := b.objects[key]; exists {
-			result[key] = make([]byte, len(data))
-			copy(result[key], data)
-		}
-	}
-
-	return result, nil
-}
-
-func (b *MockBackend) PutObjects(ctx context.Context, objects map[string][]byte) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
-
-	for key, data := range objects {
-		b.objects[key] = make([]byte, len(data))
-		copy(b.objects[key], data)
-	}
-
-	return nil
-}
-
-func (b *MockBackend) HealthCheck(ctx context.Context) error {
-	return nil
-}
-
-func (b *MockBackend) Close() error {
-	return nil
-}
-
-// TestFUSEOptimizations tests the optimized FUSE implementation
-func TestFUSEOptimizations(t *testing.T) {
-	t.Parallel()
-
-	// Create test components
-	backend := NewMockBackend()
-
-	cacheConfig := &cache.MultiLevelConfig{
-		L1Config: &cache.L1Config{
-			Enabled:    true,
-			Size:       1024 * 1024, // 1MB
-			MaxEntries: 1000,
-			TTL:        time.Minute,
-		},
-		Policy: "inclusive",
-	}
-
-	mlCache, err := cache.NewMultiLevelCache(cacheConfig)
-	require.NoError(t, err)
-
-	writeBuffer, err := vfs.NewWriter(context.Background(), backend)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = writeBuffer.Close() })
-
-	metricsConfig := &metrics.Config{
-		Enabled:   true,
-		Namespace: "test",
-	}
-
-	collector, err := metrics.NewCollector(metricsConfig)
-	require.NoError(t, err)
-
-	// Create filesystem
-	fuseConfig := &fuse.Config{
-		DefaultUID:  1000,
-		DefaultGID:  1000,
-		DefaultMode: 0644,
-	}
-
-	filesystem := fuse.NewFileSystem(t.Context(), backend, mlCache, writeBuffer, collector, fuseConfig)
-
-	// Test filesystem operations
-	_ = filesystem // Use filesystem to avoid unused variable
-	t.Run("TestReadAheadOptimization", func(t *testing.T) {
-		t.Parallel()
-
-		// Create test data in backend
-		testKey := "test-readahead.txt"
-		testData := make([]byte, 1024*1024) // 1MB of test data
-		for i := range testData {
-			testData[i] = byte(i % 256)
-		}
-		err := backend.PutObject(context.Background(), testKey, testData, nil)
-		require.NoError(t, err)
-
-		// Get filesystem stats before operations
-		statsBefore := filesystem.GetStats()
-
-		// Perform sequential reads to trigger read-ahead
-		readSize := 64 * 1024 // 64KB chunks
-		for offset := 0; offset < len(testData); offset += readSize {
-			end := min(offset+readSize, len(testData))
-
-			data, err := backend.GetObject(context.Background(), testKey, int64(offset), int64(readSize))
-			require.NoError(t, err)
-			assert.Equal(t, testData[offset:end], data)
-		}
-
-		// Check that read-ahead was triggered (or verify filesystem is functional)
-		statsAfter := filesystem.GetStats()
-		// For this test, just verify stats are being tracked
-		assert.GreaterOrEqual(t, statsAfter.Reads, statsBefore.Reads)
-	})
-
-	t.Run("TestWriteCoalescing", func(t *testing.T) {
-		t.Parallel()
-
-		testKey := "test-coalesce.txt"
-
-		// Perform multiple small writes that should be coalesced
-		writes := []struct {
-			offset int64
-			data   []byte
-		}{
-			{0, []byte("Hello, ")},
-			{7, []byte("World!")},
-			{13, []byte(" This")},
-			{18, []byte(" should")},
-			{25, []byte(" be")},
-			{28, []byte(" coalesced.")},
-		}
-
-		for _, write := range writes {
-			err := writeBuffer.Write(testKey, write.offset, write.data)
-			require.NoError(t, err)
-		}
-
-		// Flush and verify
-		err = writeBuffer.FlushAll()
-		require.NoError(t, err)
-
-		// Give time for async operations
-		time.Sleep(100 * time.Millisecond)
-
-		// Check final result
-		finalData, err := backend.GetObject(context.Background(), testKey, 0, 0)
-		require.NoError(t, err)
-
-		expected := "Hello, World! This should be coalesced."
-		assert.Equal(t, expected, string(finalData))
-	})
-
-	t.Run("TestCacheOptimization", func(t *testing.T) {
-		t.Parallel()
-
-		testKey := "test-cache.txt"
-		testData := []byte("This data should be cached for fast access")
-
-		err := backend.PutObject(context.Background(), testKey, testData, nil)
-		require.NoError(t, err)
-
-		// First read (cache miss)
-		data1, err := backend.GetObject(context.Background(), testKey, 0, 0)
-		require.NoError(t, err)
-		assert.Equal(t, testData, data1)
-
-		// Cache the data
-		mlCache.Put(testKey, 0, testData)
-
-		// Second read (should hit cache)
-		cachedData := mlCache.Get(testKey, 0, int64(len(testData)))
-		assert.Equal(t, testData, cachedData)
-
-		// Verify cache statistics
-		stats := mlCache.Stats()
-		assert.Positive(t, stats.Hits)
-	})
-
-	t.Run("TestMetricsCollection", func(t *testing.T) {
-		t.Parallel()
-
-		// Record some operations
-		collector.RecordOperation("read", 50*time.Millisecond, 1024, true)
-		collector.RecordOperation("write", 30*time.Millisecond, 2048, true)
-		collector.RecordCacheHit("test-key", 1024)
-		collector.RecordCacheMiss("another-key", 2048)
-
-		// Get metrics
-		metrics := collector.GetMetrics()
-		assert.NotNil(t, metrics)
-		assert.NotEmpty(t, metrics)
-	})
-
-	t.Run("TestPerformanceBenchmark", func(t *testing.T) {
-		t.Parallel()
-
-		if testing.Short() {
-			t.Skip("Skipping benchmark in short mode")
-		}
-
-		testKey := "bench-test.dat"
-		dataSize := 1024 * 1024 // 1MB
-		testData := make([]byte, dataSize)
-		for i := range testData {
-			testData[i] = byte(i % 256)
-		}
-
-		err := backend.PutObject(context.Background(), testKey, testData, nil)
-		require.NoError(t, err)
-
-		// Benchmark sequential reads
-		start := time.Now()
-		iterations := 10
-		totalBytes := int64(0)
-
-		for range iterations {
-			data, err := backend.GetObject(context.Background(), testKey, 0, 0)
-			require.NoError(t, err)
-			totalBytes += int64(len(data))
-		}
-
-		duration := time.Since(start)
-		throughput := float64(totalBytes) / duration.Seconds() / 1024 / 1024 // MB/s
-
-		t.Logf("Sequential read throughput: %.2f MB/s", throughput)
-		assert.Greater(t, throughput, 10.0) // Should be at least 10 MB/s
-	})
-}
-
-// TestFUSEFileOperations tests basic file operations
-func TestFUSEFileOperations(t *testing.T) {
-	t.Parallel()
-
-	backend := NewMockBackend()
-
-	cacheConfig := &cache.MultiLevelConfig{
-		L1Config: &cache.L1Config{
-			Enabled:    true,
-			Size:       512 * 1024,
-			MaxEntries: 100,
-		},
-	}
-
-	mlCache, err := cache.NewMultiLevelCache(cacheConfig)
-	require.NoError(t, err)
-
-	writeBuffer, err := vfs.NewWriter(context.Background(), backend)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = writeBuffer.Close() })
-
-	collector, err := metrics.NewCollector(&metrics.Config{Enabled: true})
-	require.NoError(t, err)
-
-	filesystem := fuse.NewFileSystem(t.Context(), backend, mlCache, writeBuffer, collector, nil)
-	_ = filesystem // Use filesystem to avoid unused variable
-
-	t.Run("FileCreationAndAccess", func(t *testing.T) {
-		t.Parallel()
-
-		// Test file creation
-		testKey := "created-file.txt"
-		testContent := []byte("This file was created through FUSE")
-
-		err := backend.PutObject(context.Background(), testKey, testContent, nil)
-		require.NoError(t, err)
-
-		// Test file reading
-		data, err := backend.GetObject(context.Background(), testKey, 0, 0)
-		require.NoError(t, err)
-		assert.Equal(t, testContent, data)
-
-		// Test partial read
-		partialData, err := backend.GetObject(context.Background(), testKey, 5, 10)
-		require.NoError(t, err)
-		assert.Equal(t, testContent[5:15], partialData)
-	})
-
-	t.Run("DirectoryOperations", func(t *testing.T) {
-		t.Parallel()
-
-		// Create directory structure
-		files := []string{
-			"dir1/file1.txt",
-			"dir1/file2.txt",
-			"dir1/subdir/file3.txt",
-			"dir2/file4.txt",
-		}
-
-		for _, file := range files {
-			content := []byte("Content of " + file)
-			err := backend.PutObject(context.Background(), file, content, nil)
-			require.NoError(t, err)
-		}
-
-		// List directory contents
-		objects, err := backend.ListObjects(context.Background(), "dir1/", 10)
-		require.NoError(t, err)
-		assert.GreaterOrEqual(t, len(objects), 2) // At least file1.txt and file2.txt
-	})
-
-	t.Run("ErrorHandling", func(t *testing.T) {
-		t.Parallel()
-
-		// Test accessing non-existent file
-		_, err := backend.GetObject(context.Background(), "non-existent.txt", 0, 0)
-		require.Error(t, err)
-
-		// Test reading beyond file bounds
-		testKey := "small-file.txt"
-		smallContent := []byte("small")
-		err = backend.PutObject(context.Background(), testKey, smallContent, nil)
-		require.NoError(t, err)
-
-		// Read beyond end of file
-		data, err := backend.GetObject(context.Background(), testKey, 100, 10)
-		require.NoError(t, err)
-		assert.Empty(t, data) // Should return empty data, not error
-	})
-}
-
-// BenchmarkFUSEOperations benchmarks FUSE performance
-func BenchmarkFUSEOperations(b *testing.B) {
-	backend := NewMockBackend()
-	mlCache, _ := cache.NewMultiLevelCache(&cache.MultiLevelConfig{
-		L1Config: &cache.L1Config{
-			Enabled:    true,
-			Size:       1024 * 1024,
-			MaxEntries: 1000,
-		},
-	})
-
-	writeBuffer, _ := vfs.NewWriter(context.Background(), backend)
-	defer func() { _ = writeBuffer.Close() }()
-
-	collector, _ := metrics.NewCollector(&metrics.Config{Enabled: false})
-	filesystem := fuse.NewFileSystem(b.Context(), backend, mlCache, writeBuffer, collector, nil)
-	_ = filesystem // Use filesystem to avoid unused variable
-
-	// Prepare test data
-	testData := make([]byte, 64*1024) // 64KB
-	for i := range testData {
-		testData[i] = byte(i % 256)
-	}
-
-	b.Run("SequentialReads", func(b *testing.B) {
-		testKey := "bench-sequential-read.dat"
-		_ = backend.PutObject(context.Background(), testKey, testData, nil)
-
-		b.ResetTimer()
-		b.ReportAllocs()
-
-		for range b.N {
-			_, err := backend.GetObject(context.Background(), testKey, 0, int64(len(testData)))
-			if err != nil {
-				b.Fatal(err)
-			}
-		}
-
-		b.SetBytes(int64(len(testData)))
-	})
-
-	b.Run("SequentialWrites", func(b *testing.B) {
-		b.ResetTimer()
-		b.ReportAllocs()
-
-		for i := range b.N {
-			testKey := "bench-write-" + string(rune(i%100)) // Cycle through 100 keys
-			err := backend.PutObject(context.Background(), testKey, testData, nil)
-			if err != nil {
-				b.Fatal(err)
-			}
-		}
-
-		b.SetBytes(int64(len(testData)))
-	})
-
-	b.Run("CachedReads", func(b *testing.B) {
-		testKey := "bench-cached-read.dat"
-		_ = backend.PutObject(context.Background(), testKey, testData, nil)
-		mlCache.Put(testKey, 0, testData) // Pre-cache the data
-
-		b.ResetTimer()
-		b.ReportAllocs()
-
-		for range b.N {
-			data := mlCache.Get(testKey, 0, int64(len(testData)))
-			if data == nil {
-				b.Fatal("Cache miss")
-			}
-		}
-
-		b.SetBytes(int64(len(testData)))
-	})
-}
